pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matrix-nio-mcp"
version = "0.1.0"
description = "MCP server for Matrix homeserver integration via matrix-nio"
requires-python = ">=3.12"
dependencies = [
    "matrix-nio[e2e]>=0.25",
    "mcp[cli]>=1.0",
    "pydantic>=2.0",
    "structlog>=24.0",
    "aiofiles>=23.0",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
matrix-nio-mcp = "matrix_nio_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/matrix_nio_mcp"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.0",
    "ruff>=0.4",
]
